# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
      "ChoreStatus": {
        "enum": [
          "PENDING",
          "COMPLETED"
        ],
        "type": "string"
      },
...
      "CreateChoreDto": {
        "properties": {
          "assigneeUserId": {
            "description": "Optional user ID to assign the chore to at creation time",
            "example": "user-uuid",
            "type": "string"
          },
          "description": {
            "example": "Take out kitchen and bathroom trash before 9pm.",
            "maxLength": 1000,
            "type": "string"
          },
          "dueDate": {
            "description": "Optional ISO-8601 due date for the chore",
            "example": "2026-03-10T21:00:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "title": {
            "description": "Short, human-readable chore title",
            "example": "Take out trash",
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "type": "object"
      },
...
      "LoginDto": {
        "properties": {
          "email": {
            "example": "alex@example.com",
            "type": "string"
          },
          "password": {
            "example": "StrongPass123!",
            "maxLength": 128,
            "minLength": 8,
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "RegisterDto": {
        "properties": {
          "email": {
            "example": "alex@example.com",
            "type": "string"
          },
          "fullName": {
            "example": "Alex Smith",
            "type": "string"
          },
          "password": {
            "example": "StrongPass123!",
            "maxLength": 128,
            "minLength": 8,
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "UpdateChoreAssigneeDto": {
        "properties": {
          "assigneeUserId": {
            "description": "User ID of the new assignee. Omit or set to null/undefined to unassign the chore completely.",
            "example": "user-uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "Supabase access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Module 1-2 platform and authentication",
    "title": "RoomieManager Backend API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns Supabase user and session."
          }
        },
        "summary": "Login with email/password via Supabase Auth",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Authenticated user identity and claims."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Return current authenticated user from access token claims",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Returns Supabase user and optional session."
          }
        },
        "summary": "Register account with email/password via Supabase Auth",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/chores/{choreId}/assign": {
      "patch": {
        "operationId": "ChoresController_updateChoreAssignee",
        "parameters": [
          {
            "in": "path",
            "name": "choreId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChoreAssigneeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns updated chore state."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign or unassign a chore.",
        "tags": [
          "Chores"
        ]
      }
    },
    "/api/v1/chores/{choreId}/complete": {
      "patch": {
        "operationId": "ChoresController_completeChore",
        "parameters": [
          {
            "in": "path",
            "name": "choreId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns updated chore state."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a chore as completed.",
        "tags": [
          "Chores"
        ]
      }
    },
...
    "/api/v1/groups/{groupId}/chores": {
      "get": {
        "operationId": "ChoresController_listGroupChores",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "1-based page number.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "example": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Maximum number of items per page.",
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "example": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Optional filter by status (PENDING or COMPLETED)",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ChoreStatus"
            }
          },
          {
            "description": "Optional filter by assignee user ID",
            "in": "query",
            "name": "assigneeUserId",
            "required": false,
            "schema": {
              "example": "user-uuid",
              "type": "string"
            }
          },
          {
            "description": "Optional lower bound for due date (inclusive)",
            "in": "query",
            "name": "dueAfter",
            "required": false,
            "schema": {
              "example": "2026-03-10T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "description": "Optional upper bound for due date (inclusive)",
            "in": "query",
            "name": "dueBefore",
            "required": false,
            "schema": {
              "example": "2026-03-20T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "description": "Optional sort field for chore listing.",
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "enum": [
                "dueDate",
                "createdAt",
                "updatedAt",
                "status"
              ],
              "type": "string"
            }
          },
          {
            "description": "Sort order applied to the selected field.",
            "in": "query",
            "name": "sortOrder",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "chores": [
                        {
                          "assignedToUserId": "550e8400-e29b-41d4-a716-446655440001",
                          "completedAt": null,
                          "createdAt": "2026-03-05T14:42:00.000Z",
                          "createdBy": "550e8400-e29b-41d4-a716-446655440001",
                          "description": null,
                          "dueDate": "2026-03-06T10:00:00.000Z",
                          "groupId": "cm8w9z0abc123def456ghi789",
                          "id": "cm8wa8qgk0004mk6z9s29u0ro",
                          "status": "PENDING",
                          "title": "Take out trash",
                          "updatedAt": "2026-03-05T14:42:00.000Z"
                        }
                      ],
                      "groupId": "cm8w9z0abc123def456ghi789",
                      "pagination": {
                        "hasNextPage": false,
                        "hasPreviousPage": false,
                        "page": 1,
                        "pageSize": 20,
                        "totalItems": 1,
                        "totalPages": 1
                      }
                    },
                    "meta": {
                      "requestId": "ac63a2cd-2f6f-4be8-86fc-0d8c91e2e1b7",
                      "timestamp": "2026-03-05T14:43:00.000Z"
                    },
                    "success": true
                  }
                }
              }
            },
            "description": "Returns chores for the specified group with pagination metadata."
          },
          "400": {
            "description": "Invalid group ID or invalid filter/pagination/sort query values."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          },
          "403": {
            "description": "Caller is not an active member of this group."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List chores for a group with optional filters.",
        "tags": [
          "Chores"
        ]
      },
      "post": {
        "operationId": "ChoresController_createChore",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChoreDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Chore created successfully."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new chore within a group.",
        "tags": [
          "Chores"
        ]
      }
    },
...
  },
  "servers": [],
...
}